requirements.txt
torch
numpy
matplotlib
opencv-python
albumentations
pillow
tqdm